refactor(runtime): move run_tool_call_loop skeleton into agent/turn — #7415
Final G1 commit. The slimmed orchestrator (loop control + the inline
steps the RUN_SHEET keeps inline) moves to turn/mod.rs, with the
condensed run sheet as its module doc. The post-hoc chunker body moves
to turn/events.rs::stream_text_posthoc_chunks (cancel keeps returning
Err(ToolLoopCancelled) via ?). MAX_MALFORMED_TOOL_PROTOCOL_RETRIES and
DEFAULT_MAX_TOOL_ITERATIONS move with the skeleton (pub(crate);
loop_.rs tests reach them via a #[cfg(test)] re-export).

loop_.rs consolidates all scattered turn re-exports into the single
ENGINE-MAP §4 block: pub use {run_tool_call_loop, StreamDelta,
DraftEvent, PROGRESS_MIN_INTERVAL_MS, ToolLoopCancelled,
is_tool_loop_cancelled, ModelSwitchRequested,
is_model_switch_requested, ModelSwitchCallback, scrub_credentials} +
#[cfg(test)] pub(crate) test-only re-exports. STREAM_CHUNK_MIN_CHARS
re-export dropped (last loop_.rs user moved with the chunker).
agent_turn/run call sites unchanged — identical loop_:: paths.

RUN_SHEET deviation (recorded there): turn/mod.rs is 624 lines, over
the 600 hard cap — the overage is module wiring with no other home
(run-sheet doc, 20 mod decls + re-exports, 30-param signature); the
cap stays binding for step files (all <=280).
