+++ crates/zeroclaw-runtime/src/agent/loop_.rs
@@ -108,7 +108,6 @@ use crate::observability::{self, Observer, ObserverEvent};
 use crate::platform;
 use crate::security::{AutonomyLevel, SecurityPolicy};
 use crate::tools::{self, Tool};
-use crate::util::truncate_with_ellipsis;
 use anyhow::{Context, Result};
 use std::collections::HashSet;
 use std::fmt::Write;
@@ -134,14 +133,6 @@ pub use super::cost::{
     TOOL_LOOP_COST_TRACKING_CONTEXT, ToolLoopCostTrackingContext, TurnUsage,
     check_tool_loop_budget, record_tool_loop_cost_usage,
 };
-
-pub(crate) use super::turn::STREAM_CHUNK_MIN_CHARS;
-/// Maximum malformed internal tool-protocol retries before returning a safe fallback.
-const MAX_MALFORMED_TOOL_PROTOCOL_RETRIES: usize = 2;
-
-/// Default maximum agentic tool-use iterations per user message to prevent runaway loops.
-/// Used as a safe fallback when `max_tool_iterations` is unset or configured as zero.
-const DEFAULT_MAX_TOOL_ITERATIONS: usize = 10;
 
 // History management moved to `super::history`.
 pub use super::history::{
@@ -412,10 +403,6 @@ fn compute_excluded_mcp_tools(
         .collect()
 }
 
-pub use super::turn::scrub_credentials;
-
-pub use super::turn::{DraftEvent, PROGRESS_MIN_INTERVAL_MS, StreamDelta};
-
 pub use zeroclaw_api::TOOL_CHOICE_OVERRIDE;
 
 /// Convert a tool registry to OpenAI function-calling format for native tool support.
@@ -563,21 +550,6 @@ pub use super::tool_execution::{
     should_execute_tools_in_parallel,
 };
 
-pub(crate) use super::turn::resolve_display_text;
-
-#[cfg(test)]
-pub(crate) use super::turn::build_native_assistant_history;
-
-pub use super::turn::{
-    ModelSwitchCallback, ModelSwitchRequested, ToolLoopCancelled, is_model_switch_requested,
-    is_tool_loop_cancelled,
-};
-
-#[cfg(test)]
-pub(crate) use super::turn::StreamedChatOutcome;
-#[cfg(test)]
-pub(crate) use super::turn::consume_provider_streaming_response;
-
 /// Execute a single turn of the agent loop: send messages, parse tool calls,
 /// execute tools, and loop until the LLM produces a final text response.
 /// When `silent` is true, suppresses stdout (for channel use).
@@ -638,541 +610,21 @@ pub async fn agent_turn(
     .await
 }
 
-#[cfg(test)]
-pub(crate) use super::turn::maybe_inject_channel_delivery_defaults;
-
 // ── Agent Tool-Call Loop ──────────────────────────────────────────────────
-// Core agentic iteration: send conversation to the LLM, parse any tool
-// calls from the response, execute them, append results to history, and
-// repeat until the LLM produces a final text-only answer.
-//
-// Loop invariant: at the start of each iteration, `history` contains the
-// full conversation so far (system prompt + user messages + prior tool
-// results). The loop exits when:
-//   • the LLM returns no tool calls (final answer), or
-//   • max_iterations is reached (runaway safety), or
-//   • the cancellation token fires (external abort).
-
-/// Append a receipt footer to the response text if any receipts were collected.
-/// Execute a single turn of the agent loop: send messages, parse tool calls,
-/// execute tools, and loop until the LLM produces a final text response.
-#[allow(clippy::too_many_arguments)]
-pub async fn run_tool_call_loop(
-    model_provider: &dyn ModelProvider,
-    history: &mut Vec<ChatMessage>,
-    tools_registry: &[Box<dyn Tool>],
-    observer: &dyn Observer,
-    provider_name: &str,
-    model: &str,
-    temperature: Option<f64>,
-    silent: bool,
-    approval: Option<&ApprovalManager>,
-    channel_name: &str,
-    channel_reply_target: Option<&str>,
-    multimodal_config: &zeroclaw_config::schema::MultimodalConfig,
-    max_tool_iterations: usize,
-    cancellation_token: Option<CancellationToken>,
-    on_delta: Option<tokio::sync::mpsc::Sender<DraftEvent>>,
-    hooks: Option<&crate::hooks::HookRunner>,
-    excluded_tools: &[String],
-    dedup_exempt_tools: &[String],
-    activated_tools: Option<&std::sync::Arc<std::sync::Mutex<crate::tools::ActivatedToolSet>>>,
-    model_switch_callback: Option<ModelSwitchCallback>,
-    pacing: &zeroclaw_config::schema::PacingConfig,
-    strict_tool_parsing: bool,
-    parallel_tools: bool,
-    max_tool_result_chars: usize,
-    context_token_budget: usize,
-    shared_budget: Option<Arc<std::sync::atomic::AtomicUsize>>,
-    channel: Option<&dyn Channel>,
-    receipt_generator: Option<&crate::agent::tool_receipts::ReceiptGenerator>,
-    collected_receipts: Option<&std::sync::Mutex<Vec<String>>>,
-) -> Result<String> {
-    let max_iterations = if max_tool_iterations == 0 {
-        DEFAULT_MAX_TOOL_ITERATIONS
-    } else {
-        max_tool_iterations
-    };
-
-    let turn_id = Uuid::new_v4().to_string();
-    let loop_started_at = Instant::now();
-    let loop_ignore_tools: HashSet<&str> = pacing
-        .loop_ignore_tools
-        .iter()
-        .map(String::as_str)
-        .collect();
-    let mut consecutive_identical_outputs: usize = 0;
-    let mut last_tool_output_hash: Option<u64> = None;
-
-    let mut loop_detector = crate::agent::loop_detector::LoopDetector::new(
-        crate::agent::loop_detector::LoopDetectorConfig {
-            enabled: pacing.loop_detection_enabled,
-            window_size: pacing.loop_detection_window_size,
-            max_repeats: pacing.loop_detection_max_repeats,
-        },
-    );
-
-    // Accumulated display text across all tool-loop calls.
-    let mut accumulated_display_text = String::new();
-    let mut malformed_tool_protocol_retries: usize = 0;
-
-    // Shared-ref context for the turn step functions. Every `&mut` the loop
-    // owns stays a loop local passed as an explicit argument (RUN_SHEET
-    // `turn.context.TurnCtx`).
-    let ctx = super::turn::TurnCtx {
-        model_provider,
-        tools_registry,
-        observer,
-        provider_name,
-        model,
-        temperature,
-        silent,
-        approval,
-        channel_name,
-        channel_reply_target,
-        multimodal_config,
-        cancellation_token: cancellation_token.as_ref(),
-        on_delta: on_delta.as_ref(),
-        hooks,
-        excluded_tools,
-        dedup_exempt_tools,
-        activated_tools,
-        pacing,
-        strict_tool_parsing,
-        parallel_tools,
-        max_tool_result_chars,
-        channel,
-        receipt_generator,
-        collected_receipts,
-        turn_id: &turn_id,
-    };
-
-    for iteration in 0..max_iterations {
-        let mut seen_tool_signatures: HashSet<(String, String)> = HashSet::new();
-
-        if cancellation_token
-            .as_ref()
-            .is_some_and(CancellationToken::is_cancelled)
-        {
-            return Err(ToolLoopCancelled.into());
-        }
-
-        // Shared iteration budget: parent + subagents share a global counter
-        if let Some(ref budget) = shared_budget {
-            let remaining = budget.load(std::sync::atomic::Ordering::Relaxed);
-            if remaining == 0 {
-                ::zeroclaw_log::record!(
-                    WARN,
-                    ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Note)
-                        .with_outcome(::zeroclaw_log::EventOutcome::Unknown)
-                        .with_attrs(::serde_json::json!({"iteration": iteration})),
-                    "Shared iteration budget exhausted at iteration "
-                );
-                break;
-            }
-            budget.fetch_sub(1, std::sync::atomic::Ordering::Relaxed);
-        }
-
-        super::turn::preflight_history_maintenance(history, context_token_budget, iteration);
-
-        // Check if model switch was requested via model_switch tool
-        if let Some(ref callback) = model_switch_callback
-            && let Ok(guard) = callback.lock()
-            && let Some((new_model_provider, new_model)) = guard.as_ref()
-            && (new_model_provider != provider_name || new_model != model)
-        {
-            ::zeroclaw_log::record!(
-                INFO,
-                ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Note),
-                &format!(
-                    "Model switch detected: {} {} -> {} {}",
-                    provider_name, model, new_model_provider, new_model
-                )
-            );
-            return Err(ModelSwitchRequested {
-                model_provider: new_model_provider.clone(),
-                model: new_model.clone(),
-            }
-            .into());
-        }
-
-        let iteration_tool_specs = super::turn::build_iteration_tool_specs(
-            model_provider,
-            tools_registry,
-            excluded_tools,
-            activated_tools,
-        );
-        let super::turn::IterationToolSpecs {
-            ref tool_specs,
-            use_native_tools,
-            ..
-        } = iteration_tool_specs;
-
-        let (vision_model_provider_box, degrade_strip_images) =
-            super::turn::resolve_vision_provider(
-                model_provider,
-                history,
-                multimodal_config,
-                provider_name,
-            )?;
-
-        let (active_model_provider, active_model_provider_name, active_model): (
-            &dyn ModelProvider,
-            &str,
-            &str,
-        ) = if let Some(ref vp_box) = vision_model_provider_box {
-            let vp_name = multimodal_config
-                .vision_model_provider
-                .as_deref()
-                .unwrap_or(provider_name);
-            let vm = multimodal_config.vision_model.as_deref().unwrap_or(model);
-            (vp_box.as_ref(), vp_name, vm)
-        } else {
-            (model_provider, provider_name, model)
-        };
-
-        let prepared_messages = super::turn::prepare_messages_for_iteration(
-            history,
-            multimodal_config,
-            degrade_strip_images,
-        )
-        .await?;
-
-        let llm_started_at = super::turn::announce_llm_request(
-            &ctx,
-            history,
-            active_model_provider,
-            active_model_provider_name,
-            active_model,
-            iteration,
-        )
-        .await;
-
-        super::turn::enforce_tool_loop_budget()?;
-
-        // Unified path via ModelProvider::chat so provider-specific native tool logic
-        // (OpenAI/Anthropic/OpenRouter/compatible adapters) is honored.
-        let request_tools = if use_native_tools {
-            Some(tool_specs.as_slice())
-        } else {
-            None
-        };
-        let should_consume_provider_stream = on_delta.is_some()
-            && model_provider.supports_streaming()
-            && (request_tools.is_none() || model_provider.supports_streaming_tool_events());
-        ::zeroclaw_log::record!(DEBUG, ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Note).with_attrs(::serde_json::json!({"has_on_delta": on_delta.is_some(), "supports_streaming": model_provider.supports_streaming(), "should_consume_provider_stream": should_consume_provider_stream})), &format!("Streaming decision for iteration {}", iteration + 1));
-
-        let super::turn::ProviderCallOutcome {
-            chat_result,
-            streamed_live_deltas,
-            streamed_protocol_suppressed,
-        } = super::turn::call_provider(
-            &ctx,
-            active_model_provider,
-            active_model,
-            &prepared_messages.messages,
-            request_tools,
-            should_consume_provider_stream,
-            iteration,
-        )
-        .await?;
-
-        let (
-            response_text,
-            parsed_text,
-            tool_calls,
-            assistant_history_content,
-            native_tool_calls,
-            parse_issue_detected,
-            protocol_suppressed,
-            response_streamed_live,
-        ) = match chat_result {
-            Ok(resp) => {
-                let interpreted = super::turn::interpret_chat_response(
-                    &ctx,
-                    resp,
-                    &iteration_tool_specs,
-                    streamed_protocol_suppressed,
-                    llm_started_at,
-                    iteration,
-                );
-                (
-                    interpreted.response_text,
-                    interpreted.parsed_text,
-                    interpreted.tool_calls,
-                    interpreted.assistant_history_content,
-                    interpreted.native_tool_calls,
-                    interpreted.parse_issue_detected,
-                    streamed_protocol_suppressed,
-                    streamed_live_deltas,
-                )
-            }
-            Err(e) => {
-                super::turn::record_llm_failure(
-                    observer,
-                    provider_name,
-                    model,
-                    llm_started_at,
-                    iteration,
-                    &turn_id,
-                    &e,
-                );
-                let recovered = super::turn::try_recover_context_overflow(history, &e, iteration);
-                if recovered {
-                    continue;
-                }
-                return Err(e);
-            }
-        };
-
-        let display_text = resolve_display_text(
-            &response_text,
-            &parsed_text,
-            !tool_calls.is_empty(),
-            !native_tool_calls.is_empty(),
-        );
-
-        // Native provider tool_calls are converted into parsed `tool_calls`
-        // above; if this branch is reached there is no valid native call to run.
-        if tool_calls.is_empty() && parse_issue_detected {
-            malformed_tool_protocol_retries += 1;
-            ::zeroclaw_log::record!(
-                WARN,
-                ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Fail)
-                    .with_outcome(::zeroclaw_log::EventOutcome::Failure)
-                    .with_attrs(serde_json::json!({
-                        "channel": channel_name,
-                        "model_provider": provider_name,
-                        "model": model,
-                        "trace_id": turn_id,
-                        "error": "malformed internal tool protocol omitted from channel output",
-                    })),
-                "tool_call_parse_feedback"
-            );
-            ::zeroclaw_log::record!(
-                DEBUG,
-                ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Note)
-                    .with_attrs(serde_json::json!({
-                    "iteration": iteration + 1,
-                    "retry": malformed_tool_protocol_retries,
-                    "max_retries": MAX_MALFORMED_TOOL_PROTOCOL_RETRIES,
-                    "response_excerpt": truncate_with_ellipsis(
-                        &scrub_credentials(&response_text),
-                        600
-                    ),
-                    })),
-                "tool_call_parse_feedback_details"
-            );
-
-            if malformed_tool_protocol_retries <= MAX_MALFORMED_TOOL_PROTOCOL_RETRIES {
-                // This is model feedback, not a tool result: malformed protocol
-                // output has no valid tool_call_id to attach a role=tool message to.
-                history.push(ChatMessage::user(
-                    "[Tool call parse error]\n\
-                     Your previous response looked like an internal tool-call protocol payload, \
-                     but ZeroClaw could not parse it into a valid tool call. Use the supported \
-                     tool-call schema, or answer in natural language if no tool is needed."
-                        .to_string(),
-                ));
-                continue;
-            }
-
-            let fallback =
-                crate::i18n::get_required_cli_string("channel-runtime-malformed-tool-output");
-            accumulated_display_text.push_str(&fallback);
-            if let Some(ref tx) = on_delta {
-                let _ = tx.send(StreamDelta::Text(fallback.to_string())).await;
-            }
-            history.push(ChatMessage::assistant(fallback.to_string()));
-            return Ok(accumulated_display_text);
-        }
-
-        // ── Progress: LLM responded ─────────────────────────────
-        if let Some(ref tx) = on_delta {
-            let llm_secs = llm_started_at.elapsed().as_secs();
-            if !tool_calls.is_empty() {
-                let _ = tx
-                    .send(StreamDelta::Status(format!(
-                        "\u{1f4ac} Got {} tool call(s) ({llm_secs}s)\n",
-                        tool_calls.len()
-                    )))
-                    .await;
-            }
-        }
-
-        if tool_calls.is_empty() {
-            ::zeroclaw_log::record!(
-                INFO,
-                ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Complete)
-                    .with_outcome(::zeroclaw_log::EventOutcome::Success)
-                    .with_attrs(::serde_json::json!({
-                        "model": model,
-                        "iteration": iteration + 1,
-                        "text": scrub_credentials(&display_text),
-                        "trace_id": turn_id,
-                    })),
-                "turn_final_response"
-            );
-            // No tool calls — this is the final response.
-            accumulated_display_text.push_str(&display_text);
-
-            // If text wasn't streamed live, send it now via post-hoc chunking.
-            // When streamed live, the channel already received the deltas.
-            if let Some(ref tx) = on_delta
-                && !response_streamed_live
-                && !protocol_suppressed
-            {
-                let mut chunk = String::new();
-                for word in display_text.split_inclusive(char::is_whitespace) {
-                    if cancellation_token
-                        .as_ref()
-                        .is_some_and(CancellationToken::is_cancelled)
-                    {
-                        return Err(ToolLoopCancelled.into());
-                    }
-                    chunk.push_str(word);
-                    if chunk.len() >= STREAM_CHUNK_MIN_CHARS
-                        && tx
-                            .send(StreamDelta::Text(std::mem::take(&mut chunk)))
-                            .await
-                            .is_err()
-                    {
-                        break;
-                    }
-                }
-                if !chunk.is_empty() {
-                    let _ = tx.send(StreamDelta::Text(chunk)).await;
-                }
-            }
-
-            history.push(ChatMessage::assistant(response_text.clone()));
-            return Ok(accumulated_display_text);
-        }
-
-        // Do not accumulate intermediate-turn display text into the final
-        // channel response. Native tool-call providers may emit narration or
-        // scratchpad-like text alongside tool calls; draft-capable channels
-        // can still see it live through `on_delta` below, but the final
-        // delivered response must only contain the final assistant turn.
-
-        // Native tool-call model_providers can return assistant text separately from
-        // the structured call payload; relay it to draft-capable channels.
-        if !display_text.is_empty() {
-            if !native_tool_calls.is_empty()
-                && let Some(ref tx) = on_delta
-            {
-                let mut narration = display_text.clone();
-                if !narration.ends_with('\n') {
-                    narration.push('\n');
-                }
-                let _ = tx.send(StreamDelta::Text(narration)).await;
-            }
-            if !silent {
-                print!("{display_text}");
-                let _ = std::io::stdout().flush();
-            }
-        }
-
-        // When multiple tool calls are present and interactive CLI approval is not needed, run
-        // tool executions concurrently for lower wall-clock latency.
-        let allow_parallel_execution =
-            parallel_tools && should_execute_tools_in_parallel(&tool_calls, approval);
-        let super::turn::PreparedToolCalls {
-            mut ordered_results,
-            executable_indices,
-            executable_calls,
-        } = super::turn::prepare_tool_calls(
-            &ctx,
-            &tool_calls,
-            &mut seen_tool_signatures,
-            iteration,
-        )
-        .await;
-
-        let executed_outcomes = if allow_parallel_execution && executable_calls.len() > 1 {
-            execute_tools_parallel(
-                &executable_calls,
-                tools_registry,
-                activated_tools,
-                observer,
-                cancellation_token.as_ref(),
-                receipt_generator,
-            )
-            .await?
-        } else {
-            execute_tools_sequential(
-                &executable_calls,
-                tools_registry,
-                activated_tools,
-                observer,
-                cancellation_token.as_ref(),
-                receipt_generator,
-            )
-            .await?
-        };
-
-        super::turn::record_executed_outcomes(
-            &ctx,
-            &executable_indices,
-            &executable_calls,
-            executed_outcomes,
-            &mut ordered_results,
-            iteration,
-        )
-        .await;
-
-        let super::turn::CollectedResults {
-            individual_results,
-            tool_results,
-            detection_relevant_output,
-        } = super::turn::collect_tool_results(
-            ordered_results,
-            &tool_calls,
-            history,
-            &mut loop_detector,
-            &loop_ignore_tools,
-            max_tool_result_chars,
-            collected_receipts,
-            model,
-            iteration,
-            &turn_id,
-        )?;
-
-        super::turn::check_identical_output_abort(
-            &detection_relevant_output,
-            loop_started_at,
-            pacing,
-            &mut consecutive_identical_outputs,
-            &mut last_tool_output_hash,
-            model,
-            iteration,
-            &turn_id,
-        )?;
-
-        super::turn::append_tool_round_to_history(
-            history,
-            assistant_history_content,
-            &native_tool_calls,
-            &individual_results,
-            &tool_results,
-            use_native_tools,
-        );
-    }
-
-    super::turn::finish_after_max_iterations(
-        model_provider,
-        history,
-        provider_name,
-        model,
-        temperature,
-        pacing,
-        cancellation_token.as_ref(),
-        max_iterations,
-        accumulated_display_text,
-        &turn_id,
-    )
-    .await
-}
+// The turn engine lives in `super::turn` — `run_tool_call_loop` plus one
+// file per step (run sheet in agent/turn/mod.rs). `crate::agent::loop_`
+// stays the canonical public path via these re-exports.
+#[cfg(test)]
+pub(crate) use super::turn::{
+    DEFAULT_MAX_TOOL_ITERATIONS, MAX_MALFORMED_TOOL_PROTOCOL_RETRIES, StreamedChatOutcome,
+    build_native_assistant_history, consume_provider_streaming_response,
+    maybe_inject_channel_delivery_defaults, resolve_display_text,
+};
+pub use super::turn::{
+    DraftEvent, ModelSwitchCallback, ModelSwitchRequested, PROGRESS_MIN_INTERVAL_MS, StreamDelta,
+    ToolLoopCancelled, is_model_switch_requested, is_tool_loop_cancelled, run_tool_call_loop,
+    scrub_credentials,
+};
 
 /// Build the tool instruction block for the system prompt so the LLM knows
 /// how to invoke tools.
